requirements.txt
tensorflow==2.3.1
textgenrnn==2.0.0
python-telegram-bot==12.8
